refactor(claw-server): remove stored profile registration paths
