+++ packages/browseros-agent/apps/claw-server/src/services/browseros-connect.ts
@@ -21,9 +21,8 @@ import {
 import { logger } from '../lib/logger'
 import { getMcpManager } from '../lib/mcp-manager'
 import { tildifyHomePath } from '../lib/tildify'
-import { type Harness, harnessEnum } from '../routes/agents/schemas'
 import { BROWSEROS_MCP_SERVER_NAME, publicMcpUrl } from '../shared/mcp-url'
-import { HARNESS_TO_AGENT_ID } from './harness-install'
+import { HARNESS_TO_AGENT_ID, type Harness, harnessEnum } from './harnesses'
 import { relinkManagedServer } from './mcp-relink'
 import { specFor } from './spec-for'
