+++ packages/browseros-agent/apps/claw-server/src/services/spec-for.ts
@@ -10,9 +10,8 @@
  * a parallel allow-list here. Stdio-only agents fall back to wrapping
  * the URL via `npx mcp-remote`, the same approach `apps/server` uses.
  *
- * Used by both profile installs (`harness-install.ts`) and the
- * single shared install (`browseros-connect.ts`) so the transport
- * rule lives in one place.
+ * Used by the shared install in `browseros-connect.ts` so the
+ * transport rule stays separate from connection orchestration.
  */
 
 import {
